tutorial 3 source code
Revised version of the source code developed for tutorial 3: (i) Included an implementation of ArrayBasedBinaryTree (and ArrayBasedBinaryTreNode). In this, there is an alternative printing of the tree which visualizes the levels of the tree, and also prints the underlying array. (ii) Moved all the exercises (computeHeight, computeOddNumbers, isProperBinaryTree, computeOddNumbersBetweenLevels) in a class BinaryTreeUtils.